Move all custom types to separate sub-package.
