+++ roll35/settlement.py
@@ -9,18 +9,18 @@ import random
 
 from nextcord.ext import commands
 
-from .cog import Cog
 from .common import bad_return
 from .data.constants import RANK
 from .magicitem import roll_many
 from .retcode import Ret
+from .types.cog import R35Cog
 
 NOT_READY = 'Settlement data is not yet available, please try again later.'
 
 logger = logging.getLogger(__name__)
 
 
-class Settlement(Cog):
+class Settlement(R35Cog):
     @commands.command()
     async def settlement(self, ctx, population):
         '''Roll magic items for a settlement with the given population.'''
